delete function now works

+++ server/db/migrations/20241011164337_create_post_tags_table.js
@@ -6,7 +6,7 @@ exports.up = function (knex) {
     return knex.schema.createTable('post_tags', (table) => {
         table.increments('id').primary();
         table.integer('post_id').notNullable();
-        table.foreign('post_id').references('id').inTable('posts');
+        table.foreign('post_id').references('id').inTable('posts').onDelete('CASCADE');;
         table.integer('tag_id').notNullable();
         table.foreign('tag_id').references('id').inTable('tags');
     });
@@ -17,6 +17,9 @@ exports.up = function (knex) {
  * @returns { Promise<void> }
  */
 exports.down = function (knex) {
-    return knex.schema.dropTable('post_tags');
-
+    return knex.schema.hasTable('post_tags').then((exists) => {
+        if (exists) {
+            return knex.schema.dropTable('post_tags');
+        }
+    });
 };
